cli:get:set: migrate tests to (quicker) nosetests3
